refactor: extract function
